Moved add_source_description_to_source_column to ascwds utils file
