+++ utils/ind_cqc_filled_posts_utils/ascwds_filled_posts_calculator/calculate_ascwds_filled_posts_return_worker_record_count_if_equal_to_total_staff.py
@@ -4,11 +4,9 @@ from utils.column_values.categorical_column_values import (
     ASCWDSFilledPostsSource as Source,
 )
 from utils.ind_cqc_filled_posts_utils.ascwds_filled_posts_calculator.utils import (
+    add_source_description_to_source_column,
     ascwds_filled_posts_is_null,
     two_cols_are_equal_and_at_least_minimum_permitted_value,
-)
-from utils.ind_cqc_filled_posts_utils.utils import (
-    add_source_description_to_source_column,
 )
 
 ascwds_filled_posts_totalstaff_equal_wkrrecs_source_description = (
