+++ utils/ind_cqc_filled_posts_utils/ascwds_filled_posts_calculator/utils.py
@@ -1,4 +1,4 @@
-import pyspark.sql.functions as F
+from pyspark.sql import DataFrame, functions as F
 
 from utils.ind_cqc_filled_posts_utils.ascwds_filled_posts_calculator.calculation_constants import (
     ASCWDSFilledPostCalculationConstants as calculation_constant,
@@ -61,3 +61,37 @@ def absolute_difference_between_two_columns(first_col: str, second_col: str) -> 
 
 def average_of_two_columns(first_col: str, second_col: str):
     return (F.col(first_col) + F.col(second_col)) / 2
+
+
+def add_source_description_to_source_column(
+    input_df: DataFrame,
+    populated_column_name: str,
+    source_column_name: str,
+    source_description: str,
+) -> DataFrame:
+    """
+    Adds a source description to a specified column if the populated column is not null and the source column is null.
+
+    This function is run following a calculation that populates a column with data.
+    If the populated column has a value and the source column is null, it updates the source column with the provided source description.
+    This is useful for tracking the origin of the data in the source column.
+
+    Args:
+        input_df (DataFrame): The input Spark DataFrame.
+        populated_column_name (str): Name of the column used to determine if the row should have a source description.
+        source_column_name (str): Name of the column to update with the source description.
+        source_description (str): The description to insert into the source column.
+
+    Returns:
+        DataFrame: A new DataFrame with the source column updated.
+    """
+    return input_df.withColumn(
+        source_column_name,
+        F.when(
+            (
+                F.col(populated_column_name).isNotNull()
+                & F.col(source_column_name).isNull()
+            ),
+            source_description,
+        ).otherwise(F.col(source_column_name)),
+    )
